+++ webui/module/Media/src/Media/Controller/MediaController.php
@@ -27,7 +27,6 @@ namespace Media\Controller;
 
 use Zend\Mvc\Controller\AbstractActionController;
 use Zend\View\Model\ViewModel;
-use Zend\Json\Json;
 use Exception;
 
 class MediaController extends AbstractActionController
@@ -135,72 +134,6 @@ class MediaController extends AbstractActionController
         return new ViewModel(array(
             'volume' => $volumename,
         ));
-    }
-
-    /**
-     * Get Data Action
-     *
-     * @return object
-     */
-    public function getDataAction()
-    {
-        $this->RequestURIPlugin()->setRequestURI();
-
-        if (!$this->SessionTimeoutPlugin()->isValid()) {
-            return $this->redirect()->toRoute(
-                'auth',
-                array(
-                    'action' => 'login'
-                ),
-                array(
-                    'query' => array(
-                        'req' => $this->RequestURIPlugin()->getRequestURI(),
-                        'dird' => $_SESSION['bareos']['director']
-                    )
-                )
-            );
-        }
-
-        $result = null;
-
-        $data = $this->params()->fromQuery('data');
-        $volume = $this->params()->fromQuery('volume');
-
-        if ($data == "all") {
-            try {
-                $this->bsock = $this->getServiceLocator()->get('director');
-                $result = $this->getMediaModel()->getVolumes($this->bsock);
-                $this->bsock->disconnect();
-            } catch (Exception $e) {
-                echo $e->getMessage();
-            }
-        } elseif ($data == "details") {
-            try {
-                $this->bsock = $this->getServiceLocator()->get('director');
-                // workaround until llist volume returns array instead of object
-                $result[0] = $this->getMediaModel()->getVolume($this->bsock, $volume);
-                $this->bsock->disconnect();
-            } catch (Exception $e) {
-                echo $e->getMessage();
-            }
-        } elseif ($data == "jobs") {
-            try {
-                $this->bsock = $this->getServiceLocator()->get('director');
-                $result = $this->getMediaModel()->getVolumeJobs($this->bsock, $volume);
-                $this->bsock->disconnect();
-            } catch (Exception $e) {
-                echo $e->getMessage();
-            }
-        }
-
-        $response = $this->getResponse();
-        $response->getHeaders()->addHeaderLine('Content-Type', 'application/json');
-
-        if (isset($result)) {
-            $response->setContent(JSON::encode($result));
-        }
-
-        return $response;
     }
 
     /**