+++ CHANGELOG.md
@@ -57,6 +57,7 @@ and since Bareos version 20 this project adheres to [Semantic Versioning](https:
 - dird: deprecate client `Autoprune`, `JobRetention`, and `FileRetention` [PR #1425]
 - daemons: remove deprecated `Pid Directory` config option, and update `Maximum Concurrent Jobs` default value to 1 [PR #1426]
 - build: switch to FreeBSD 12.4 [PR #1440]
+- webui: create internal api module [PR #1447]
 
 ### Removed
 - remove no longer used pkglists [PR #1335]
@@ -142,4 +143,5 @@ and since Bareos version 20 this project adheres to [Semantic Versioning](https:
 [PR #1429]: https://github.com/bareos/bareos/pull/1429
 [PR #1439]: https://github.com/bareos/bareos/pull/1439
 [PR #1440]: https://github.com/bareos/bareos/pull/1440
+[PR #1447]: https://github.com/bareos/bareos/pull/1447
 [unreleased]: https://github.com/bareos/bareos/tree/master